+++ RN/components/Modal/GenerateModal.tsx
@@ -1,14 +1,10 @@
-import TextInput from "@/components/TextInput";
 import { BaseModal, BaseModalProps } from "@/components/Modal/BaseModal";
 import Button from "@/components/Button/Button";
-import { StyleSheet, Text, View } from "react-native";
-import React, { useState } from "react";
-import { generateWifiQRCodeString } from "@/helpers/generateWifiQRCodeString";
-
-export interface OnGeneratePressArgs {
-  name: string;
-  url: string;
-}
+import { StyleSheet, View } from "react-native";
+import { useState } from "react";
+import { URLModalContent } from "@/components/Modal/URLModalContent";
+import { WifiModalContent } from "@/components/Modal/WifiModalContent";
+import { OnGeneratePressArgs } from "@/types";
 
 export interface GenerateModalProps
   extends Pick<BaseModalProps, "isVisible" | "onClose"> {
@@ -21,45 +17,13 @@ export function GenerateModal({
   onClose,
 }: GenerateModalProps) {
   const [type, setType] = useState<"URL" | "Wifi" | null>(null);
-  const [name, setName] = React.useState("");
-  const [url, setURL] = React.useState("");
-
-  const [ssid, setSSID] = useState("");
-  const [password, setPassword] = useState("");
-  const [encryption, setEncryption] = useState("");
 
   const changeType = (type: "URL" | "Wifi") => {
     setType(type);
   };
 
-  function handleURLGenerate() {
-    onGeneratePress?.({
-      name,
-      url,
-    });
-
-    setName("");
-    setURL("");
-  }
-
-  function handleWifiGenerate() {
-    if (encryption !== "WPA" && encryption !== "WEP") {
-      return;
-    }
-
-    onGeneratePress?.({
-      name,
-      url: generateWifiQRCodeString({
-        encryption,
-        ssid,
-        password,
-      }),
-    });
-
-    setName("");
-    setSSID("");
-    setPassword("");
-    setEncryption("");
+  function handleGenerate(args: OnGeneratePressArgs) {
+    onGeneratePress?.(args);
   }
 
   function handleClose() {
@@ -96,86 +60,22 @@ export function GenerateModal({
     </>
   );
 
-  const URLContent = (
-    <>
-      <TextInput
-        label="Name"
-        value={name}
-        placeholder="Instagram"
-        onChangeText={setName}
-      />
-      <View style={styles.spacer8} />
-      <TextInput
-        label="URL"
-        value={url}
-        placeholder="https://www.instagram.com/<your_id>/"
-        onChangeText={setURL}
-      />
-
-      <View style={styles.spacer32} />
-
-      <Button
-        title="Generate"
-        buttonType="primary"
-        state="default"
-        onPress={handleURLGenerate}
-      />
-      <View style={styles.spacer8} />
-      <Button
-        title="Cancel"
-        buttonType="link"
-        state="default"
-        onPress={handleClose}
-      />
-    </>
-  );
-
-  const WifiContent = (
-    <>
-      <TextInput
-        label="Name"
-        value={name}
-        placeholder="Wifi"
-        onChangeText={setName}
-      />
-      <View style={styles.spacer8} />
-      <TextInput label="SSID" value={ssid} onChangeText={setSSID} />
-      <View style={styles.spacer8} />
-      <TextInput label="Password" value={password} onChangeText={setPassword} />
-      <View style={styles.spacer8} />
-      <TextInput
-        label="Encryption"
-        value={encryption}
-        placeholder="WPA | WEP"
-        onChangeText={setEncryption}
-      />
-
-      <View style={styles.spacer32} />
-
-      <Button
-        title="Generate"
-        buttonType="primary"
-        state="default"
-        onPress={handleWifiGenerate}
-      />
-      <View style={styles.spacer8} />
-      <Button
-        title="Cancel"
-        buttonType="link"
-        state="default"
-        onPress={handleClose}
-      />
-    </>
-  );
-
   return (
     <BaseModal isVisible={isVisible} onClose={handleClose}>
       <View>
-        {type === "URL"
-          ? URLContent
-          : type === "Wifi"
-            ? WifiContent
-            : generalContent}
+        {type === "URL" ? (
+          <URLModalContent
+            onGeneratePress={handleGenerate}
+            onCancel={handleClose}
+          />
+        ) : type === "Wifi" ? (
+          <WifiModalContent
+            onGeneratePress={handleGenerate}
+            onCancel={handleClose}
+          />
+        ) : (
+          generalContent
+        )}
       </View>
     </BaseModal>
   );
