+++ RN/types.ts
@@ -2,3 +2,8 @@ export type HomeListItemType = {
   title: string;
   url: string;
 };
+
+export interface OnGeneratePressArgs {
+  name: string;
+  url: string;
+}
